'before_change_category'

+++ server/src/controllers/userController.js
@@ -107,36 +107,74 @@ const createUser = async (req, res, next) => {
 
 const updateUser = async (req, res, next) => {
     const { id } = req.params;
+    const { userupdate, auth } = req.headers;
     const { email, useName, password, status, role } = req.body;
-    try {
-        await User.update(
-            {
-                email: email,
-                user_name: useName,
-                password: password,
-                status: status,
-            },
-            {
-                where: {
-                    id: id,
-                },
-            }
-        );
 
-        await User_role.update(
-            {
-                RoleId: role,
-            },
-            {
-                where: {
-                    UserId: id,
-                },
+    try {
+        //If user update are 'admin'
+        if (userupdate === "admin") {
+            //If user login are 'admin'
+            if (auth === "admin") {
+                //If user update not change name and status
+                if (userupdate === useName && status === "1") {
+                    await User.update(
+                        {
+                            email: email,
+                            password: password,
+                        },
+                        {
+                            where: {
+                                id: id,
+                            },
+                        }
+                    );
+                    res.send({
+                        message: "Update user success",
+                        action: "update",
+                    });
+                } else {
+                    res.send({
+                        message:
+                            "Update fail, this account only can update email and password",
+                        action: "warning",
+                    });
+                }
+            } else {
+                res.send({
+                    message: "You are not authentication",
+                    action: "warning",
+                });
             }
-        );
-        res.send({
-            message: "Update user success",
-            action: "update",
-        });
+        } else {
+            await User.update(
+                {
+                    email: email,
+                    user_name: useName,
+                    password: password,
+                    status: status,
+                },
+                {
+                    where: {
+                        id: id,
+                    },
+                }
+            );
+
+            await User_role.update(
+                {
+                    RoleId: role,
+                },
+                {
+                    where: {
+                        UserId: id,
+                    },
+                }
+            );
+            res.send({
+                message: "Update user success",
+                action: "update",
+            });
+        }
     } catch (error) {
         console.log(error);
     }
@@ -144,22 +182,30 @@ const updateUser = async (req, res, next) => {
 
 const deleteUser = async (req, res, next) => {
     const { id } = req.params;
+    const { name } = req.headers;
     try {
-        await User.destroy({
-            where: {
-                id: id,
-            },
-        });
+        if (name === "admin") {
+            res.send({
+                message: "You cannot delete this user",
+                action: "warning",
+            });
+        } else {
+            await User.destroy({
+                where: {
+                    id: id,
+                },
+            });
 
-        await User_role.destroy({
-            where: {
-                UserId: id,
-            },
-        });
-        res.send({
-            message: "Delete user success",
-            action: "delete",
-        });
+            await User_role.destroy({
+                where: {
+                    UserId: id,
+                },
+            });
+            res.send({
+                message: "Delete user success",
+                action: "delete",
+            });
+        }
     } catch (error) {
         console.log(error);
     }
